feat: update logic
