+++ ui/app/(espace-pro)/layout.tsx
@@ -1,28 +1,13 @@
-import { headers } from "next/headers"
 import type { PropsWithChildren } from "react"
-import { type IUserRecruteurPublic } from "shared"
+
+import { getSession } from "@/utils/getSession"
 
 import { Footer } from "../_components/Footer"
 
 import { EspaceProHeader } from "./_components/EspaceProHeader"
 
-export async function getSession(): Promise<IUserRecruteurPublic | null> {
-  try {
-    const headerStore = await headers()
-    const userRaw = headerStore.get("x-user")
-
-    if (!userRaw) {
-      return null
-    }
-
-    return JSON.parse(userRaw)
-  } catch (error) {
-    return null
-  }
-}
-
 export default async function RecruteurLayout({ children }: PropsWithChildren) {
-  const user = await getSession()
+  const { user } = await getSession()
 
   return (
     <>
